added boss animations (still some animaitons missing???) and they work, now the camera lock indicator works properly. and fixxed an issue where enemies couldnt take damage

+++ Scrunglys - Staat 1 Periode 4/Assets/Scripts/Can/Movement/Movement.cs
@@ -12,8 +12,8 @@ public class Movement : MonoBehaviour
     public CheckCurrentSword currentSword;
 
     [Header("Basic movement")]
-    public int walkSpeed;
-    public int sprintSpeed;
+    public float walkSpeed;
+    public float sprintSpeed;
     public float gravity;
     public float sprintStamina;
